ongoing refactoring (3)
