--- src/templates/ops.js
@@ -1,1 +1,0 @@
-store.ops.js
